python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "pandas>=3.0.2",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]